Move controllers and concerns under app/controllers and app/concerns
In order to get files automatically reloaded in Rails, these files should be under `ActiveSupport::Dependencies.autoload_paths` directory. `Rails::Engine` adds `app/controllers` path automatically to `autoload_paths`.

Reference: https://github.com/rails/rails/issues/12195#issuecomment-24432724

Closes #697
