+++ app/concerns/active_admin/controllers/resource/action_builder.rb
@@ -1,0 +1,22 @@
+module ActiveAdmin
+  module Controllers
+    module Resource
+
+      module ActionBuilder
+        extend ActiveSupport::Concern
+
+        module ClassMethods
+
+          def clear_member_actions!
+            active_admin_config.clear_member_actions!
+          end
+
+          def clear_collection_actions!
+            active_admin_config.clear_collection_actions!
+          end
+        end
+      end
+
+    end
+  end
+end
