+++ app/concerns/active_admin/controllers/resource/sidebars.rb
@@ -1,0 +1,19 @@
+module ActiveAdmin
+  module Controllers
+    module Resource
+
+      module Sidebars
+        protected
+
+        def skip_sidebar!
+          @skip_sidebar = true
+        end
+
+        def skip_sidebar?
+          @skip_sidebar == true
+        end
+      end
+
+    end
+  end
+end
