+++ app/controllers/active_admin/devise/confirmations_controller.rb
@@ -1,0 +1,7 @@
+module ActiveAdmin
+  module Devise
+    class ConfirmationsController < ::Devise::ConfirmationsController
+      include ::ActiveAdmin::Devise::Controller
+    end
+  end
+end
