+++ app/controllers/active_admin/devise/passwords_controller.rb
@@ -1,0 +1,7 @@
+module ActiveAdmin
+  module Devise
+    class PasswordsController < ::Devise::PasswordsController
+      include ::ActiveAdmin::Devise::Controller
+    end
+  end
+end
