+++ app/controllers/active_admin/devise/registrations_controller.rb
@@ -1,0 +1,7 @@
+module ActiveAdmin
+  module Devise
+    class RegistrationsController < ::Devise::RegistrationsController
+      include ::ActiveAdmin::Devise::Controller
+    end
+  end
+end
