+++ app/controllers/active_admin/devise/sessions_controller.rb
@@ -1,0 +1,7 @@
+module ActiveAdmin
+  module Devise
+    class SessionsController < ::Devise::SessionsController
+      include ::ActiveAdmin::Devise::Controller
+    end
+  end
+end
