+++ app/controllers/active_admin/devise/unlocks_controller.rb
@@ -1,0 +1,7 @@
+module ActiveAdmin
+  module Devise
+    class UnlocksController < ::Devise::UnlocksController
+      include ::ActiveAdmin::Devise::Controller
+    end
+  end
+end
