+++ app/controllers/active_admin/page_controller.rb
