+++ app/controllers/active_admin/resource_controller.rb
@@ -1,26 +1,28 @@
 require 'inherited_resources'
-require 'active_admin/resource_controller/action_builder'
-require 'active_admin/resource_controller/data_access'
-require 'active_admin/resource_controller/decorators'
-require 'active_admin/resource_controller/scoping'
-require 'active_admin/resource_controller/sidebars'
-require 'active_admin/resource_controller/resource_class_methods'
+
+require 'active_admin/controllers/resource/action_builder'
+require 'active_admin/controllers/resource/decorators'
+require 'active_admin/controllers/resource/data_access'
+require 'active_admin/controllers/resource/scoping'
+require 'active_admin/controllers/resource/sidebars'
+require 'active_admin/controllers/resource/resource_class_methods'
 
 module ActiveAdmin
   # All Resources Controller inherits from this controller.
   # It implements actions and helpers for resources.
   class ResourceController < BaseController
+
     layout :determine_active_admin_layout
 
     respond_to :html, :xml, :json
     respond_to :csv, :only => :index
 
-    include ActionBuilder
-    include Decorators
-    include DataAccess
-    include Scoping
-    include Sidebars
-    extend  ResourceClassMethods
+    include Controllers::Resource::ActionBuilder
+    include Controllers::Resource::Decorators
+    include Controllers::Resource::DataAccess
+    include Controllers::Resource::Scoping
+    include Controllers::Resource::Sidebars
+    extend  Controllers::Resource::ResourceClassMethods
 
     def self.active_admin_config=(config)
       if @active_admin_config = config
