+++ lib/active_admin.rb
@@ -17,8 +17,8 @@ require 'arbre'
 require 'active_admin/helpers/i18n'
 
 module ActiveAdmin
-
   autoload :VERSION,                  'active_admin/version'
+  autoload :AccessDenied,             'active_admin/access_denied'
   autoload :Application,              'active_admin/application'
   autoload :AssetRegistration,        'active_admin/asset_registration'
   autoload :Authorization,            'active_admin/authorization_adapter'
@@ -26,7 +26,6 @@ module ActiveAdmin
   autoload :Breadcrumbs,              'active_admin/breadcrumbs'
   autoload :Callbacks,                'active_admin/callbacks'
   autoload :Component,                'active_admin/component'
-  autoload :BaseController,           'active_admin/base_controller'
   autoload :ControllerAction,         'active_admin/controller_action'
   autoload :CSVBuilder,               'active_admin/csv_builder'
   autoload :Dependencies,             'active_admin/dependencies'
@@ -44,11 +43,9 @@ module ActiveAdmin
   autoload :OrderClause,              'active_admin/order_clause'
   autoload :Page,                     'active_admin/page'
   autoload :PagePresenter,            'active_admin/page_presenter'
-  autoload :PageController,           'active_admin/page_controller'
   autoload :PageDSL,                  'active_admin/page_dsl'
   autoload :Reloader,                 'active_admin/reloader'
   autoload :Resource,                 'active_admin/resource'
-  autoload :ResourceController,       'active_admin/resource_controller'
   autoload :ResourceDSL,              'active_admin/resource_dsl'
   autoload :Scope,                    'active_admin/scope'
   autoload :ScopeChain,               'active_admin/helpers/scope_chain'
