+++ lib/active_admin/devise.rb
@@ -55,26 +55,5 @@ module ActiveAdmin
         prefix + path
       end
     end
-
-    class SessionsController < ::Devise::SessionsController
-      include ::ActiveAdmin::Devise::Controller
-    end
-
-    class PasswordsController < ::Devise::PasswordsController
-      include ::ActiveAdmin::Devise::Controller
-    end
-
-    class UnlocksController < ::Devise::UnlocksController
-      include ::ActiveAdmin::Devise::Controller
-    end
-
-    class RegistrationsController < ::Devise::RegistrationsController
-       include ::ActiveAdmin::Devise::Controller
-    end
-    
-    class ConfirmationsController < ::Devise::ConfirmationsController
-       include ::ActiveAdmin::Devise::Controller
-    end
-
   end
 end
